Move runProc + related classes into src/tool/run_proc.dart.
