+++ lib/src/report/dependencies.dart
@@ -18,7 +18,7 @@ import '../logging.dart';
 import '../model.dart';
 import '../package_context.dart';
 import '../pubspec.dart';
-import '../sdk_env.dart';
+import '../tool/run_proc.dart';
 import '../utils.dart';
 
 import '_common.dart';
